+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # master
 *Please put new entries at the top.
 
+1. KVO observations can now be made with Smart Key Path in Swift 3.2+, using `producer(for:)` and `signal(for:)` available on `NSObject.reactive`. (#3491, kudos to @andersio)
+
+
 # 6.0.1
+
 1. [Xcode 9 beta 5] Fixed an issue causing infinite recursion in the Swift runtime. (#3498, kudos to @andersio)
 
 # 6.0.0